+++ platform/apple/LinkPortableUds.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 /* LINK-owned Apple UDS amalgamation for product faces. */
+#include "../../src/crypto/aes_cmac.c"
 #include "../../src/uds/uds.c"
 #include "../../src/uds/uds_services.c"
 #include "../../src/uds/uds_server.c"